+++ actionpack/lib/action_dispatch/system_testing/driver.rb
@@ -3,11 +3,12 @@
 module ActionDispatch
   module SystemTesting
     class Driver # :nodoc:
-      def initialize(name, **options)
+      def initialize(name, **options, &desired_capabilities)
         @name = name
         @browser = Browser.new(options[:using])
         @screen_size = options[:screen_size]
         @options = options[:options]
+        @desired_capabilities = desired_capabilities
       end
 
       def use
@@ -22,6 +23,8 @@ module ActionDispatch
         end
 
         def register
+          define_browser_capabilities(@browser.driver_option)
+
           Capybara.register_driver @name do |app|
             case @name
             when :selenium then register_selenium(app)
@@ -29,6 +32,10 @@ module ActionDispatch
             when :webkit then register_webkit(app)
             end
           end
+        end
+
+        def define_browser_capabilities(driver_option)
+          @desired_capabilities.call(driver_option) if @desired_capabilities
         end
 
         def browser_options