+++ actionpack/test/dispatch/system_testing/driver_test.rb
@@ -2,6 +2,7 @@
 
 require "abstract_unit"
 require "action_dispatch/system_testing/driver"
+require "selenium/webdriver"
 
 class DriverTest < ActiveSupport::TestCase
   test "initializing the driver" do
@@ -50,5 +51,71 @@ class DriverTest < ActiveSupport::TestCase
 
   test "registerable? returns false if driver is rack_test" do
     assert_not ActionDispatch::SystemTesting::Driver.new(:rack_test).send(:registerable?)
+  end
+
+  test "define extra capabilities using chrome" do
+    driver_option = nil
+    driver = ActionDispatch::SystemTesting::Driver.new(:selenium, screen_size: [1400, 1400], using: :chrome) do |option|
+      option.add_argument("start-maximized")
+      option.add_emulation(device_name: "iphone 6")
+      option.add_preference(:detach, true)
+
+      driver_option = option
+    end
+    driver.use
+
+    expected = { args: ["start-maximized"], mobileEmulation: { deviceName: "iphone 6" }, prefs: { detach: true } }
+    assert_equal expected, driver_option.as_json
+  end
+
+  test "define extra capabilities using headless_chrome" do
+    driver_option = nil
+    driver = ActionDispatch::SystemTesting::Driver.new(:selenium, screen_size: [1400, 1400], using: :headless_chrome) do |option|
+      option.add_argument("start-maximized")
+      option.add_emulation(device_name: "iphone 6")
+      option.add_preference(:detach, true)
+
+      driver_option = option
+    end
+    driver.use
+
+    expected = { args: ["start-maximized"], mobileEmulation: { deviceName: "iphone 6" }, prefs: { detach: true } }
+    assert_equal expected, driver_option.as_json
+  end
+
+  test "define extra capabilities using firefox" do
+    driver_option = nil
+    driver = ActionDispatch::SystemTesting::Driver.new(:selenium, screen_size: [1400, 1400], using: :firefox) do |option|
+      option.add_preference("browser.startup.homepage", "http://www.seleniumhq.com/")
+      option.add_argument("--host=127.0.0.1")
+
+      driver_option = option
+    end
+    driver.use
+
+    expected = { "moz:firefoxOptions" => { args: ["--host=127.0.0.1"], prefs: { "browser.startup.homepage" => "http://www.seleniumhq.com/" } } }
+    assert_equal expected, driver_option.as_json
+  end
+
+  test "define extra capabilities using headless_firefox" do
+    driver_option = nil
+    driver = ActionDispatch::SystemTesting::Driver.new(:selenium, screen_size: [1400, 1400], using: :headless_firefox) do |option|
+      option.add_preference("browser.startup.homepage", "http://www.seleniumhq.com/")
+      option.add_argument("--host=127.0.0.1")
+
+      driver_option = option
+    end
+    driver.use
+
+    expected = { "moz:firefoxOptions" => { args: ["--host=127.0.0.1"], prefs: { "browser.startup.homepage" => "http://www.seleniumhq.com/" } } }
+    assert_equal expected, driver_option.as_json
+  end
+
+  test "does not define extra capabilities" do
+    driver = ActionDispatch::SystemTesting::Driver.new(:selenium, screen_size: [1400, 1400], using: :firefox)
+
+    assert_nothing_raised do
+      driver.use
+    end
   end
 end